+++ frontend/src/components/host/SpaceInfoUpdateBox/useSpaceUpdateForm.ts
@@ -1,4 +1,5 @@
 import { AxiosError } from 'axios';
+import { useState } from 'react';
 import { useMutation } from 'react-query';
 import { useNavigate } from 'react-router-dom';
 
@@ -11,7 +12,16 @@ import { ID } from '@/types';
 const useSpaceUpdateForm = () => {
   const navigate = useNavigate();
 
+  const [isActiveSubmit, setIsActiveSubmit] = useState(true);
+
   const { openToast } = useToast();
+
+  const onChangeSpaceName = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const input = e.target as HTMLInputElement;
+
+    const isExistValue = input.value.length > 0;
+    setIsActiveSubmit(isExistValue);
+  };
 
   const { mutate: updateSpace } = useMutation(
     ({ spaceId, name, imageUrl }: { spaceId: ID; name: string; imageUrl: string }) =>
@@ -35,7 +45,7 @@ const useSpaceUpdateForm = () => {
     updateSpace({ spaceId, name, imageUrl });
   };
 
-  return { onSubmitUpdateSpace };
+  return { isActiveSubmit, onChangeSpaceName, onSubmitUpdateSpace };
 };
 
 export default useSpaceUpdateForm;
